move localhost storage injection script to own util file for better maintenance
