+++ ui/desktop/src/main.ts
@@ -36,6 +36,7 @@ import {
   updateEnvironmentVariables,
   updateSchedulingEngineEnvironment,
 } from './utils/settings';
+import localStorageInjectionScript from './utils/localStorageInjectionScript';
 import * as crypto from 'crypto';
 // import electron from "electron";
 import * as yaml from 'yaml';
@@ -664,52 +665,10 @@ const createChat = async (
     // 2. This happens during window initialization before the renderer's React app is ready
     // 3. The timing is critical - we need to set config before React components try to read it
     // 4. Direct IPC communication would require the renderer to be fully loaded first
-    mainWindow.webContents
-      .executeJavaScript(
-        `
-      (function() {
-        let retryCount = 0;
-        const maxRetries = 5;
-        const baseDelay = 100;
-        
-        function setConfig() {
-          try {
-            if (window.localStorage && typeof window.localStorage.setItem === 'function') {
-              localStorage.setItem('gooseConfig', '${configStr}');
-              console.log('[Renderer] Successfully set localStorage config');
-              return true;
-            } else {
-              console.warn('[Renderer] localStorage not available or setItem not a function');
-            }
-          } catch (e) {
-            console.warn('[Renderer] localStorage access failed:', e);
-          }
-          return false;
-        }
-
-        function retrySetConfig() {
-          if (setConfig()) {
-            return; // Success, no need to retry
-          }
-          
-          retryCount++;
-          if (retryCount < maxRetries) {
-            const delay = baseDelay * Math.pow(2, retryCount - 1); // Exponential backoff
-            console.log(\`[Renderer] Retrying localStorage config set (attempt \${retryCount + 1}/\${maxRetries}) in \${delay}ms\`);
-            setTimeout(retrySetConfig, delay);
-          } else {
-            console.error('[Renderer] Failed to set localStorage after all retries - continuing without localStorage config');           
-          }
-        }
-
-        // Initial attempt
-        retrySetConfig();
-      })();
-    `
-      )
-      .catch((error) => {
-        console.error('Failed to execute localStorage script:', error);
-      });
+    const injectionScript = localStorageInjectionScript('gooseConfig', configStr);
+    mainWindow.webContents.executeJavaScript(injectionScript).catch((error) => {
+      console.error('Failed to execute localStorage script:', error);
+    });
   });
 
   // Handle new window creation for links
